nlp_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2

# LATER INSTALLED FOR BACKFILL.PY
psycopg2-binary
requests psycopg2-binary